Move files for creation of new target catalogs.
