+++ src/hipscat_import/control.py
@@ -1,8 +1,8 @@
 """Flow control and scripting entry points."""
 import sys
 
-import hipscat_import.run_import as runner
-from hipscat_import.command_line_arguments import parse_command_line
+import hipscat_import.catalog.run_import as runner
+from hipscat_import.catalog.command_line_arguments import parse_command_line
 
 
 def main():
